requirements.txt
# Core
fastapi 
uvicorn 
python-dotenv 
pydantic 

# AI & LLM - DeepSeek
openai 
langchain 
langchain-openai 
langchain-community 

# Vector Database - Pinecone
pinecone 
pinecone-datasets 

# Document Processing
pypdf 
python-docx 

# Text Processing
tiktoken 
httpx   

# Frontend (served via FastAPI)
jinja2 

# Integration (for Hootsuite/Sprout Social)
requests 
oauthlib 


# Utilities
python-multipart 
aiofiles 


unstructured
python-pptx
openpyxl
pandas
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2


# pip install unstructured
# pip install python-pptx
# pip install openpyxl
# pip install pandas
# pip install google-api-python-client
# pip install google-auth
# pip install google-auth-oauthlib
# pip install google-auth-httplib2

# loader = CampaignDocumentLoader("data")

# docs = loader.load_google_drive("YOUR_GOOGLE_DRIVE_FOLDER_ID")